moved pitch and swing descriptions to utils
