+++ web/src/components/playing/PlayView.tsx
@@ -55,40 +55,6 @@ export const verticalLocations = {
   4: "Low Ball",
 };
 
-export const getPitchDescription = (s: number, h: number, v: number) => {
-  const isStrike = h === 0 || h === 4 || v === 4 || v === 0 ? "A ball" : "A strike";
-  const speed = s === 0 ? "Fast" : "Slow";
-  let point = "";
-  if (v < 2) {
-    point = h < 2 ? ", high and inside" : h === 2 ? " and high" : ", high and outside";
-  }
-  if (v === 2) {
-    point = h < 2 ? " and inside" : h === 2 ? " and down the middle" : " and outside";
-  }
-  if (v > 2) {
-    point = h < 2 ? ", low and inside" : h === 2 ? " and low" : ", low and outside";
-  }
-  return `${isStrike}: ${speed}${point}.`;
-};
-
-export const getSwingDescription = (k: number, h: number, v: number) => {
-  if (k === 2) {
-    return "Nope. You are taking the pitch.";
-  }
-  const kind = k === 0 ? "For contact" : "For power";
-  let point = "";
-  if (v < 2) {
-    point = h < 2 ? "high and inside" : h === 2 ? "high" : "high and outside";
-  }
-  if (v === 2) {
-    point = h < 2 ? "inside" : h === 2 ? "down the middle" : "outside";
-  }
-  if (v > 2) {
-    point = h < 2 ? "low and inside" : h === 2 ? "low" : "low and outside";
-  }
-  return `${kind}; ${point}.`;
-};
-
 export const pitchSpeed = {
   0: "Fast",
   1: "Slow",
