+++ web/src/utils/messages.ts
@@ -44,3 +44,37 @@ export const progressMessage = (session: Session) => {
     return `Strike three! ${session.pair.pitcher?.name} wins it for the visitors `;
   }
 };
+
+export const getPitchDescription = (s: number, h: number, v: number) => {
+  const isStrike = h === 0 || h === 4 || v === 4 || v === 0 ? "A ball" : "A strike";
+  const speed = s === 0 ? "Fast" : "Slow";
+  let point = "";
+  if (v < 2) {
+    point = h < 2 ? ", high and inside" : h === 2 ? " and high" : ", high and outside";
+  }
+  if (v === 2) {
+    point = h < 2 ? " and inside" : h === 2 ? " and down the middle" : " and outside";
+  }
+  if (v > 2) {
+    point = h < 2 ? ", low and inside" : h === 2 ? " and low" : ", low and outside";
+  }
+  return `${isStrike}: ${speed}${point}.`;
+};
+
+export const getSwingDescription = (k: number, h: number, v: number) => {
+  if (k === 2) {
+    return "Nope. You are taking the pitch.";
+  }
+  const kind = k === 0 ? "For contact" : "For power";
+  let point = "";
+  if (v < 2) {
+    point = h < 2 ? "high and inside" : h === 2 ? "high" : "high and outside";
+  }
+  if (v === 2) {
+    point = h < 2 ? "inside" : h === 2 ? "down the middle" : "outside";
+  }
+  if (v > 2) {
+    point = h < 2 ? "low and inside" : h === 2 ? "low" : "low and outside";
+  }
+  return `${kind}; ${point}.`;
+};
